+++ src/js/index.js
@@ -1,3 +1,3 @@
-import App from "./App.js";
+import App from './components/App.js';
 
-new App(document.querySelector("#app"));
+new App(document.querySelector('#app'));
